WIP: Begin codebase reorg via class templates
Note that the code currently does not compile due to an "undefined
symbols" error. After much head scratching and head banging, I have
realized this might be due to the fact that Process pointers are being
stored in the readyQueue priority queue, but I have yet to confirm this
as a full solution.